+++ README.md
@@ -1,4 +1,4 @@
-# Unified Planning Bridge
+# Embedded Systems Bridge
 
 General functionalities for using unified-planning in applications.
