+++ docs/callable_interface.md
@@ -1,13 +1,13 @@
-  The callable interface of `up_bridge`
-=========================================
+  The callable interface of the `esb`
+=======================================
 
 Assuming you have fluents and actions definitions in your application domain,
-there are various ways to create UP representations of them using the `up_bridge`.
+there are various ways to create UP representations of them using the `esb`.
 The best way depends on what format your input parameter has and whether you want
 to associate the executable implementation of your callable at the same time
 when you are creating its UP representation. If your fluents or actions can have
 different implementations which you want to choose later, you need to call
-the `up_bridge` methods correctly to only declare a fluent or action.
+the `esb` methods correctly to only declare a fluent or action.
 
 The most general format for your input parameter is a `dict` describing its parameters
 and return value. You can always use this option of `create_action()` by using the
