+++ esb/bridge.py
@@ -46,7 +46,7 @@ from unified_planning.plans import (
 )
 from unified_planning.shortcuts import BoolType, IntType, OneshotPlanner, RealType, UserType
 
-from up_bridge.components.graph import plan_to_dependency_graph
+from esb.components.graph import plan_to_dependency_graph
 
 
 class Bridge:
