+++ examples/time_triggered.py
@@ -22,8 +22,8 @@ import unified_planning as up
 from unified_planning.model import EndTiming, StartTiming
 from unified_planning.shortcuts import ClosedTimeInterval, Not, OneshotPlanner
 
-from up_bridge.bridge import Bridge
-from up_bridge.plexmo import PlanDispatcher
+from esb.bridge import Bridge
+from esb.plexmo import PlanDispatcher
 
 
 #################### 1. Define the domain ####################
